Move this directory to the new location for SwingsetTests
