+++ test/swingsetTests/presenceCorruption/vat-bob.js
